refactor: parachain observers
